Release Candidate 2 (0.6.1) Fixes several issues related to support for complex types. More complex type unit tests.

+++ src/main/java/org/ic4j/candid/Bytes.java
@@ -37,7 +37,7 @@ final public class Bytes {
 		return new Bytes(input);
 	}
 
-	public Integer leb128Read() {
+	public Long leb128Read() {
 		try
 		{
 			return Leb128.readUnsigned(this.data);
